requirements.txt
# ReportHub

Professional Report Card Management System with AI-powered analytics and career assessment.

## Features
- Automated report card generation
- AI-powered performance analysis
- Career path recommendations
- M-PESA payment integration
- School branding customization

## Tech Stack
- Backend: Python/Flask
- Database: SQLite/PostgreSQL
- Frontend: HTML/CSS/JavaScript
- AI: OpenAI GPT-4
- Payment: M-PESA API

## Setup Instructions
1. Clone repository
2. Install dependencies: pip install -r requirements.txt
3. Set environment variables
4. Run: python app.py
"@ | Out-File -FilePath README.md -Encoding utf8

# 6. Create requirements.txt
@"
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-login==0.6.2
flask-migrate==4.0.5
pandas==2.0.3
numpy==1.24.3
openai==0.28.0
stripe==5.5.0
gunicorn==21.2.0
python-dotenv==1.0.0
werkzeug==2.3.7
requests==2.31.0
scikit-learn==1.3.0
openpyxl==3.1.2
pillow==10.0.0
pytesseract==0.3.10
psycopg2-binary==2.9.7
redis==5.0.0
